demo/requirements.txt
google-generativeai
python-dotenv
pathlib
python-box
streamlit
arxiv